[x/ts] - refactored primitive package, added test cases, and improved telem primitive encoding/decoding

+++ x/go/telem/series_factory.go
@@ -38,8 +38,8 @@ func NewSeriesV[T Sample](data ...T) Series { return NewSeries(data) }
 
 // MakeSeries allocates a new Series with the specified DataType and length. Note that
 // this function allocates a length and not a capacity.
-func MakeSeries(dt DataType, len int64) Series {
-	return Series{DataType: dt, Data: make([]byte, len*int64(dt.Density()))}
+func MakeSeries(dt DataType, len int) Series {
+	return Series{DataType: dt, Data: make([]byte, len*int(dt.Density()))}
 }
 
 // NewSeriesSecondsTSV creates a new Series containing TimeStamp values. All input timestamps
